updated word count, more standardization/cleanup
